GodofGods -> God of fire!
the god of gods has been replaced with god of fire, who has his own custom abilities and weapon!

The god of fire is still the most powerfull god, and his era is the age of ash along with the earth god (the earth god's status is stronger however)

The Staff Of  Fire: the god of fire's weapon, which shoots the same projectiles as mage's, and has the chaos laser in it, which deals a huge amount of damage

the god of fire also can summon red tornado's which set things on fire, fire storm, clouds of ash, and can summon demons, fire skulls and mages. he also can explode fire from himself

the god of fire deals alot of damage while also having weak armor and damage

ive tried balancing him but hes too op, pls balance things out for me!