STORMS
ive created a new ability, storm! it uses the same animation has antimatter bombs and allows you to have a object that does a action while having the antimatter animation!

ive replaced the cloudofdarkness ability with creating a storm that is big, creates lighting, forces and curses

ive also made the god of fire able to create a fire storm that moves around while heating up the place